+++ src/Plainion.GraphViz.Modules.CodeInspection/Properties/AssemblyInfo.cs
@@ -5,15 +5,15 @@ using System.Reflection;
 [assembly: AssemblyProductAttribute("Plainion.GraphViz")]
 [assembly: AssemblyDescriptionAttribute("Plainion.GraphViz")]
 [assembly: AssemblyCopyrightAttribute("Copyright @ 2017")]
-[assembly: AssemblyVersionAttribute("2.3.0")]
-[assembly: AssemblyFileVersionAttribute("2.3.0")]
+[assembly: AssemblyVersionAttribute("2.4.0")]
+[assembly: AssemblyFileVersionAttribute("2.4.0")]
 namespace System {
     internal static class AssemblyVersionInformation {
         internal const System.String AssemblyTitle = "Plainion.GraphViz.Modules.CodeInspection";
         internal const System.String AssemblyProduct = "Plainion.GraphViz";
         internal const System.String AssemblyDescription = "Plainion.GraphViz";
         internal const System.String AssemblyCopyright = "Copyright @ 2017";
-        internal const System.String AssemblyVersion = "2.3.0";
-        internal const System.String AssemblyFileVersion = "2.3.0";
+        internal const System.String AssemblyVersion = "2.4.0";
+        internal const System.String AssemblyFileVersion = "2.4.0";
     }
 }